third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/navigation-abort-detaches-frame-expected.txt

CONSOLE MESSAGE: line 18: PASS: Expected 'abort', got 'abort'.
PASS: but only if nothing crashed.
